refactor: package structure 2nd reformed
