+++ src/main/java/shop/project/pathorderserver/store/StoreRequest.java
@@ -9,4 +9,5 @@ public class StoreRequest {
     // TODO: 매장 메뉴옵션 등록
     // TODO: 매장 메뉴옵션 수정
     // TODO: 매장 메뉴옵션 삭제
+    // TODO: 주문 수정 (매장 측: 주문 상태 변경)
 }
