+++ src/main/java/shop/project/pathorderserver/user/UserController.java
@@ -5,8 +5,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import shop.project.pathorderserver._core.utils.ApiUtil;
-import shop.project.pathorderserver.order.OrderRequest;
-import shop.project.pathorderserver.order.OrderResponse;
 import shop.project.pathorderserver.order.OrderService;
 
 import java.io.IOException;
@@ -40,8 +38,8 @@ public class UserController {
     }
 
     @PostMapping("/api/users/{userId}/orders") // 주문하기
-    public ResponseEntity<?> order(@RequestBody OrderRequest.OrderDTO reqDTO) {
-        OrderResponse.OrderDTO respDTO = orderService.createOrder(reqDTO);
+    public ResponseEntity<?> order(@RequestBody UserRequest.OrderDTO reqDTO) {
+        UserResponse.OrderDTO respDTO = orderService.createOrder(reqDTO);
 
         return ResponseEntity.ok(new ApiUtil(respDTO));
     }
