+++ src/main/java/shop/project/pathorderserver/user/UserRequest.java
@@ -2,6 +2,8 @@ package shop.project.pathorderserver.user;
 
 import lombok.Data;
 
+import java.util.List;
+
 public class UserRequest {
 
     @Data // 회원 가입
@@ -37,5 +39,30 @@ public class UserRequest {
     public static class ImgDTO {
         private String username;
         private String encodedImg;
+    }
+
+
+    @Data
+    public static class OrderDTO {
+        private int storeId;
+        private String storeName;
+        private int customerId;
+        private String customerNickname;
+        private String request;
+        private List<OrderMenuDTO> orderMenuList;
+
+        @Data
+        public static class OrderMenuDTO {
+            private int menuId;
+            private String name;
+            private int price;
+            private List<OrderMenuOptionDTO> orderMenuOptionList;
+        }
+
+        @Data
+        public static class OrderMenuOptionDTO {
+            private String name;
+            private int price;
+        }
     }
 }
