+++ src/main/java/shop/project/pathorderserver/user/UserResponse.java
@@ -9,6 +9,7 @@ import shop.project.pathorderserver.order.OrderMenuOption;
 import shop.project.pathorderserver.order.OrderStatus;
 
 import java.sql.Timestamp;
+import java.util.ArrayList;
 import java.util.List;
 
 public class UserResponse {
@@ -209,6 +210,64 @@ public class UserResponse {
 
         public String getTotalPrice() { // 19500 -> 19,500 변환
             return FormatUtil.decimalFormatter(totalPrice);
+        }
+    }
+    @Data
+    public static class OrderDTO {
+        private int storeId;
+        private String storeName;
+        private int customerId;
+        private String customerNickname;
+        private String request;
+        private int totalPrice;
+        private OrderStatus status;
+        private List<OrderMenuDTO> orderMenuList;
+
+        public OrderDTO(Order order, List<OrderMenu> orderMenus, List<OrderMenuOption> orderMenuOptions) {
+            this.storeId = order.getStore().getId();
+            this.storeName = order.getStoreName();
+            this.customerId = order.getCustomer().getId();
+            this.customerNickname = order.getCustomerNickname();
+            this.request = order.getRequest();
+            this.totalPrice = order.getTotalPrice();
+            this.status = order.getStatus();
+            this.orderMenuList = orderMenus.stream()
+                    .map(orderMenu -> new OrderMenuDTO(orderMenu, orderMenuOptions))
+                    .toList();
+        }
+
+        @Data
+        public static class OrderMenuDTO {
+            private int id;
+            // private int menuId;
+            private String name;
+            private int qty;
+            private int price;
+            private List<OrderMenuOptionDTO> orderMenuOptionList = new ArrayList<>();
+
+            public OrderMenuDTO(OrderMenu orderMenu, List<OrderMenuOption> orderMenuOptions) {
+                this.id = orderMenu.getId();
+                this.name = orderMenu.getName();
+                this.price = orderMenu.getPrice();
+                for (int i = 0; i < orderMenuOptions.size(); i++) {
+                    OrderMenuOptionDTO orderMenuOption = new OrderMenuOptionDTO(orderMenuOptions.get(i));
+                    if (id == orderMenuOptions.get(i).getOrderMenu().getId()) {
+                        // this.orderMenuOptionList = orderMenuOptions.stream().map(OrderMenuOptionDTO::new).toList();
+                        this.orderMenuOptionList.add(orderMenuOption);
+                    }
+                }
+            }
+
+            @Data
+            public static class OrderMenuOptionDTO {
+                private String name;
+                private int price;
+
+                public OrderMenuOptionDTO(OrderMenuOption orderMenuOption) {
+                    this.name = orderMenuOption.getName();
+                    this.price = orderMenuOption.getPrice();
+                }
+            }
         }
     }
 }
